refactor: move Prize from constant package into domain package
