+++ src/main/java/lotto/domain/LottoStatistics.java
@@ -2,7 +2,6 @@ package lotto.domain;
 
 import lombok.Builder;
 import lombok.Getter;
-import lotto.constant.Prize;
 
 public class LottoStatistics {
