+++ src/test/java/lotto/domain/PrizeTest.java
@@ -1,5 +1,6 @@
-package lotto.constant;
+package lotto.domain;
 
+import lotto.domain.Prize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
